requirements.txt
fastapi
python-multipart
pillow
torch
torchvision
numpy
pandas
filetype
uvicorn